# documentation/openapi.yaml
openapi: 3.0.3
info:
  title: FileLoading API
  description: >-
    File Loading API - loads and processes network files (converted from ntfileload.4gl).
    Supports multiple file classes: CDR, CHG, EBL, SVC, and more.
    Includes file management, transfer operations, parser configuration, and workflow tracking.
  version: "4.0.0"
  contact:
    name: Select Software
servers:
  - url: http://localhost:5140
    description: Local development
  - url: https://api.example.com
    description: Production
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /api/v4/file-loading/load:
    post:
      operationId: post_api_v4_file_loading_load
      tags:
        - File Loading
      summary: Load a network file for processing
      description: Load a network file for processing by specifying a file path and type.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoadFileRequest"
      responses:
        "202":
          description: File load accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileLoadResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                Error: "Invalid file type code"
                ErrorCode: "INVALID_FILE_TYPE"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v4/file-loading/upload:
    post:
      operationId: post_api_v4_file_loading_upload
      tags:
        - File Loading
      summary: Upload and load a file
      description: Upload a file via multipart form data and load it for processing.
      parameters:
        - name: fileType
          in: query
          required: true
          description: File type code
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
              required:
                - file
      responses:
        "202":
          description: File upload accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileLoadResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                Error: "File is empty"
                ErrorCode: "EMPTY_FILE"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v4/file-loading/files/{nt-file-num}:
    get:
      operationId: get_api_v4_file_loading_files_nt_file_num
      tags:
        - File Loading
      summary: Get file load status
      description: Get file load status by nt-file-num (the database record key assigned at load time).
      parameters:
        - $ref: "#/components/parameters/NtFileNum"
      responses:
        "200":
          description: File status retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                Error: "File not found"
                ErrorCode: "NOT_FOUND"
  /api/v4/file-loading/files:
    get:
      operationId: get_api_v4_file_loading_files
      tags:
        - File Loading
      summary: List loaded files with filtering
      description: List loaded files with optional filtering by file type code, customer number, and max records.
      parameters:
        - name: fileTypeCode
          in: query
          required: false
          description: Filter by file type code (e.g. CDR, CHG, SSSWHLSCDR)
          schema:
            type: string
        - name: ntCustNum
          in: query
          required: false
          description: Filter by customer number
          schema:
            type: string
        - name: maxRecords
          in: query
          required: false
          description: Maximum records to return (default 100)
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: File list retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileListResponse"
        "204":
          description: No files found
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v4/file-loading/file-types:
    get:
      operationId: get_api_v4_file_loading_file_types
      tags:
        - File Loading
      summary: Get supported file types for loading
      description: Returns the list of supported file types for loading operations.
      responses:
        "200":
          description: File type list retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileTypeListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v4/file-loading/files/{nt-file-num}/reprocess:
    post:
      operationId: post_api_v4_file_loading_files_nt_file_num_reprocess
      tags:
        - File Loading
      summary: Reprocess a previously loaded file
      description: Reprocess a previously loaded file by its nt-file-num.
      parameters:
        - $ref: "#/components/parameters/NtFileNum"
      responses:
        "202":
          description: Reprocess accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileLoadResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                Error: "File not found"
                ErrorCode: "NOT_FOUND"

  # ==================================================
  # File Management - Dashboard
  # ==================================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication via X-API-Key header

  # ================================================
  # Reusable Parameters
  # ================================================
  parameters:
    NtFileNum:
      name: nt-file-num
      in: path
      required: true
      description: File number (nt_file_num) - the database record key assigned at load time
      schema:
        type: integer
...
  responses:
    Unauthorized:
      description: Authentication required - missing or invalid JWT token or API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            Error: "Unauthorized"
            ErrorCode: "UNAUTHORIZED"

  # ================================================
  # Schemas
  # ================================================
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        Error:
          type: string
          description: Human-readable error message
          example: "An error occurred"
        ErrorCode:
          type: string
          description: Machine-readable error code
          example: "INVALID_REQUEST"
        Details:
          description: Additional error details (optional)
          nullable: true

    # --- File Loading Request/Response ---
    LoadFileRequest:
      type: object
      required:
        - FileName
        - FileType
      properties:
        FileName:
          type: string
          description: Full path to file
          example: "/data/files/CDR_20260320.csv"
        FileType:
          type: string
          description: File type code (from file_type table)
          example: TEL_GSM
        NtCustNum:
          type: string
          nullable: true
          description: Network customer number
        FileDate:
          type: string
          format: date-time
          nullable: true
          description: File date (defaults to today)
        ExpectedSequence:
          type: integer
          nullable: true
          description: Expected sequence number (for validation)
    FileLoadResponse:
      type: object
      properties:
        NtFileNum:
          type: integer
          description: File number (nt_file_num)
          example: 12345
        FileName:
          type: string
          description: File name
          example: "CDR_20260320.csv"
        FileType:
          type: string
          description: File type code
          example: TEL_GSM
        Status:
          type: string
          description: Status description
          example: Pending
        StatusId:
          type: integer
          description: Status ID
          example: 1
        RecordsLoaded:
          type: integer
          description: Records loaded count
          example: 0
        RecordsFailed:
          type: integer
          description: Records failed count
          example: 0
        StartedAt:
          type: string
          format: date-time
          nullable: true
          description: Processing start time
        CompletedAt:
          type: string
          format: date-time
          nullable: true
          description: Processing completion time
    FileStatusResponse:
      type: object
      properties:
        NtFileNum:
          type: integer
          description: File number (nt_file_num)
          example: 12345
        FileName:
          type: string
          description: File name
          example: "CDR_20260320.csv"
        FileType:
          type: string
          description: File type code
          example: TEL_GSM
        NtCustNum:
          type: string
          description: Network customer number
          example: "CUST001"
        NtFileSeq:
          type: integer
          description: File sequence number
          example: 42
        StatusId:
          type: integer
          description: Status ID
          example: 2
        StatusDescription:
          type: string
          description: Status description
          example: "Transactions loaded"
        NtFileDate:
          type: string
          format: date-time
          nullable: true
          description: File date
        CreatedTm:
          type: string
          format: date-time
          nullable: true
          description: Created timestamp
        TotalRecords:
          type: integer
          nullable: true
          description: Total records in file
          example: 5000
        TotalCost:
          type: number
          format: decimal
          nullable: true
          description: Total cost
          example: 12345.67
        EarliestCall:
          type: string
          format: date-time
          nullable: true
          description: Earliest call date in file
        LatestCall:
          type: string
          format: date-time
          nullable: true
          description: Latest call date in file
    FileListResponse:
      type: object
      properties:
        Items:
          type: array
          items:
            $ref: "#/components/schemas/FileStatusResponse"
        TotalCount:
          type: integer
          description: Total number of matching files
    FileTypeListResponse:
      type: object
      properties:
        Items:
          type: array
          items:
            $ref: "#/components/schemas/FileTypeInfo"
    FileTypeInfo:
      type: object
      properties:
        Code:
          type: string
          description: File type code
          example: TEL_GSM
        Description:
          type: string
          description: File type description
          example: "Telstra GSM CDRs"
        FileClassCode:
          type: string
          description: File class code (CDR, CHG, etc.)
          example: CDR
        FileClassDescription:
          type: string
          description: File class description
          example: "Call Detail Records"
        NtCustNum:
          type: string
          description: Network customer number
          example: "CUST001"
        NtFileName:
          type: string
          description: File name pattern
          example: "TEL_GSM_*.csv"
        SkipHdr:
          type: integer
          nullable: true
          description: Skip header records count
        SkipTlr:
          type: integer
          nullable: true
          description: Skip trailer records count

    # --- File Management ---
...